Final commit on Episode CRUD, Done and tested

+++ aura/app/controllers/EpisodioController.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import models.*;
 import org.hibernate.Hibernate;
 import org.json.JSONArray;
+import org.json.JSONException;
 import org.json.simple.JSONObject;
 import play.db.jpa.JPA;
 import play.db.jpa.Transactional;
@@ -51,7 +52,7 @@ public class EpisodioController extends Controller {
 
             } catch (Exception e) {
                 e.printStackTrace();
-                return null;
+                return Results.badRequest("Error al crear el episodio");
             }
 
            //getNotification(idPaciente, intensidad, horasSuenio, regularidad, localizacion, estres);
@@ -81,15 +82,25 @@ public class EpisodioController extends Controller {
 
     @Transactional
     public static Result get(Long id) {
-        Episodio p = JPA.em().getReference(Episodio.class, id);
+        Episodio e = JPA.em().find(Episodio.class, id);
+        if(e == null)
+            return Results.ok("El episodio no existe");
+        e = JPA.em().getReference(Episodio.class, id);
         Hibernate.initialize(Episodio.class);
-        return Results.ok(Json.toJson(p));
+        return Results.ok(Json.toJson(e));
     }
 
     @Transactional
     public static Result delete(Long id1, Long id2){
+        Paciente pac = JPA.em().find(Paciente.class, id1);
+        System.out.println(pac);
+        if(pac == null)
+            return Results.ok("El paciente no existe");
         Episodio p = JPA.em().find(Episodio.class, id2);
+        if(p == null)
+            return Results.ok("El episodio no existe");
         if(p.getPacienteID().equals(id1)) {
+            p = JPA.em().getReference(Episodio.class, id2);
             JPA.em().remove(p);
             return Results.ok();
 
@@ -101,9 +112,10 @@ public class EpisodioController extends Controller {
     @BodyParser.Of(BodyParser.Json.class)
     public static Result update(Long id1, Long id2) {
 
-        Paciente p = JPA.em().getReference(Paciente.class, id1);
+        Paciente p = JPA.em().find(Paciente.class, id1);
 
         if(p != null) {
+            p = JPA.em().getReference(Paciente.class, id1);
             JsonNode j = Controller.request().body().asJson();
             System.out.println(j);
 
@@ -117,7 +129,14 @@ public class EpisodioController extends Controller {
 
             try {
 
-                Episodio e = JPA.em().getReference(Episodio.class, id2);
+                Episodio e = JPA.em().find(Episodio.class, id2);
+                if(e == null)
+                    return Results.ok("El episodio no existe");
+
+                if(!e.getPacienteID().equals(id1))
+                    return Results.ok("El paciente no tiene ese episodio");
+
+                e = JPA.em().getReference(Episodio.class, id2);
                 e.setEpisodioEstreCercano(estres);
                 e.setIdUrl(url);
                 e.setIntensidad(intensidad);
@@ -128,7 +147,7 @@ public class EpisodioController extends Controller {
 
             } catch (Exception e) {
                 e.printStackTrace();
-                return Results.ok("Error al crear el episodio");
+                return Results.ok("Error al actualizar el episodio");
             }
 
             return Results.created();
@@ -139,8 +158,8 @@ public class EpisodioController extends Controller {
     }
 
     @Transactional
-    public static Collection<Episodio> getPerDatesAnalisis(Long idP, String f1, String f2)
-    {
+    public static Collection<Episodio> getPerDatesAnalisis(Long idP, String f1, String f2) {
+
         Date d1 = parseDate(f1);
         Date d2 = parseDate(f2);
         Query query = JPA.em().createQuery("SELECT e FROM Episodio e WHERE e.fechaPublicacion >= :d1 AND e.fechaPublicacion <= :d2");
@@ -153,13 +172,19 @@ public class EpisodioController extends Controller {
     @Transactional
     public static Result getPerDates(Long idP, String f1, String f2) {
 
+        Paciente p = JPA.em().find(Paciente.class, idP);
+        if(p == null)
+            return Results.ok("El paciente no existe");
+
         Date d1 = parseDate(f1);
         Date d2 = parseDate(f2);
         Query query = JPA.em().createQuery("SELECT e FROM Episodio e WHERE e.fechaPublicacion >= :d1 AND e.fechaPublicacion <= :d2");
         query.setParameter("d1", d1);
         query.setParameter("d2", d2);
         Collection<Episodio> episodios = query.getResultList();
+        System.out.println(episodios);
         return Results.ok(Json.toJson(episodios));
+
     }
 
     private static Date parseDate(String representation) {
@@ -454,41 +479,39 @@ public class EpisodioController extends Controller {
 
 
     public static JSONObject getNotification(Long idP,int intensidad, int horasSuenio, boolean suenioRegular, int lugar, boolean episodioEstreCercano) {
-        JSONArray not = new JSONArray();
-            String inten="",suenio="",est="";
-            String mensaje = "Debe tener en cuenta las siguientes consideraciones:";
-            if(intensidad>7){
-             inten="La intensidad de su dolor es muy fuerte. Considere acudir al médico";
-            }
-            else if(intensidad>4&&intensidad<=7){
-                 inten="La intensidad de su dolor es estandar. Procure reposar para evitar que empeore.";
-            }
-            else{
-                 inten="La intensidad de su dolor es suave.";
-            }
-            if(horasSuenio>8){
-                 suenio="Sus horas de sueño son adecuadas y probablemente no sean la causa da sus migrañas";
-            }
-        else if(horasSuenio>5&&horasSuenio<=8){
-                 suenio="Sus horas de sueño son suficientes, pero debería considerar dormir un poco más.";
-            }
-        else{
-                 suenio="Sus horas de sueño son muy bajas y causan que tenga migraña.";
-            }
-        if(episodioEstreCercano){
+
+        String inten="",suenio="",est="";
+        String mensaje = "Debe tener en cuenta las siguientes consideraciones:";
+        if(intensidad > 7){
+            inten="La intensidad de su dolor es muy fuerte. Considere acudir al médico";
+        }
+        else if(intensidad > 4 && intensidad <= 7) {
+            inten = "La intensidad de su dolor es estandar. Procure reposar para evitar que empeore.";
+        }
+        else {
+            inten = "La intensidad de su dolor es suave.";
+        }
+        if(horasSuenio > 8){
+            suenio = "Sus horas de sueño son adecuadas y probablemente no sean la causa da sus migrañas";
+        }
+        else if(horasSuenio > 5 && horasSuenio <= 8) {
+            suenio="Sus horas de sueño son suficientes, pero debería considerar dormir un poco más.";
+        }
+        else {
+            suenio="Sus horas de sueño son muy bajas y causan que tenga migraña.";
+        }
+        if(episodioEstreCercano) {
              est="El estres produce dolores de cabeza muy fuertes, y es posible que esto le cause migrañas";
         }
-        else{
+        else {
              est="";
         }
 
         JSONObject simple = new JSONObject();
         simple.put("mensaje",mensaje);
-        simple.put("inten",inten);
+        simple.put("intensidad",inten);
         simple.put("suenio",suenio);
-        simple.put("est",est);
-
-            not.put(simple);
+        simple.put("estres",est);
         return simple;
     }
 
@@ -550,6 +573,92 @@ public class EpisodioController extends Controller {
     @BodyParser.Of(BodyParser.Json.class)
     public static Result updateMedicine(Long idp, Long id1, Long id2) {
         return Results.TODO;
+    }
+
+    // EPISODIO COMPLETO
+
+    @Transactional
+    @BodyParser.Of(BodyParser.Json.class)
+    public static Result createEpisode(Long paciente) throws JSONException {
+
+        JsonNode j = Controller.request().body().asJson();
+
+        Long idUrl = j.findPath("idUrl").asLong();
+        int intensidad = j.findPath("intensidad").asInt();
+        int horasSuenio = j.findPath("horasSuenio").asInt();
+        boolean regular = j.findPath("regular").asBoolean();
+        int localizacion = j.findPath("lugar").asInt();
+        boolean estres = j.findPath("estres").asBoolean();
+        String fecha = j.findPath("fechaPublicacion").asText();
+        Date thisDate = new Date();
+
+        if(!fecha.equals(""))
+            thisDate = parseDate(fecha);
+
+        Paciente estePaciente = JPA.em().find(Paciente.class, paciente);
+        System.out.println(estePaciente);
+        if(estePaciente == null)
+            return Results.ok("El paciente no existe");
+
+        Episodio e = Episodio.create(idUrl, thisDate, intensidad, horasSuenio, regular, localizacion, estres, paciente);
+        JPA.em().persist(e);
+
+        List<JsonNode> g = j.findValues("sintomas");
+
+        if(g.size() > 0) {
+            JsonNode values = g.get(0);
+            for(JsonNode js : values) {
+                int sintomaAux = js.findPath("sintoma").asInt();
+                Sintoma s = new Sintoma();
+                s.setEpisodioId(e.getId());
+                s.setSintoma(sintomaAux);
+                JPA.em().persist(s);
+            }
+        }
+
+        List<JsonNode> m = j.findValues("medicamentos");
+
+        if(m.size() > 0) {
+            JsonNode values = m.get(0);
+            for(JsonNode js : values) {
+                String nombre = js.findPath("nombre").asText();
+                int horas = js.findPath("horas").asInt();
+                Medicamento mAxux = Medicamento.create(nombre, horas, e.getId());
+                JPA.em().persist(mAxux);
+            }
+        }
+
+        List<JsonNode> med = j.findValues("actividades");
+
+        if(med.size() > 0) {
+            JsonNode values = med.get(0);
+            for(JsonNode js : values) {
+                int descripcion = js.findPath("descripcion").asInt();
+                int inte = js.findPath("intensidad").asInt();
+                int lugar = js.findPath("lugar").asInt();
+                int clima = js.findPath("clima").asInt();
+                boolean hidratacion = js.findPath("hidratacion").asBoolean();
+                ActividadFisica af = ActividadFisica.create(descripcion, inte, lugar, clima, hidratacion, e.getId());
+                JPA.em().persist(af);
+            }
+        }
+
+        List<JsonNode> al = j.findValues("alimentos");
+
+        if(al.size() > 0) {
+            JsonNode values = al.get(0);
+            for(JsonNode js : values) {
+                String nombre = js.findPath("nombre").asText();
+                int cant = js.findPath("cantidad").asInt();
+
+                Alimento all = Alimento.create(nombre, cant, e.getId());
+                JPA.em().persist(all);
+
+            }
+        }
+
+        return Results.ok(Json.toJson(EpisodioController.getNotification(paciente, intensidad, horasSuenio, regular, localizacion, estres)));
+        //return Results.ok();
     }
 
     @Transactional
